Откореектировано на прием чисел с точкой

+++ Task1/Program.cs
@@ -1,10 +1,10 @@
 ﻿// Напишите программу, которая на 
 //вход принимает два числа и выдаёт, какое число большее, а какое меньшее.
 
-int a , b;
+double a , b;
 Console.WriteLine("Введите сравниваемые числа");
-a =Convert.ToInt32(Console.ReadLine());
-b =Convert.ToInt32(Console.ReadLine());
+a =Convert.ToDouble(Console.ReadLine());
+b =Convert.ToDouble(Console.ReadLine());
 
 if (a > b)
 {
